requirements.txt
google-generativeai
langchain
langchain-google-genai
chromadb
typer
python-dotenv

